+++ docs/CHANGELOG.md
@@ -1,5 +1,6 @@
 Changes after fork:
 
+* Added a reverse SOCKS5 proxy over ssh. Configure in `configs/ssh.json`
 * Uses tmux as a pseudo-C2-like interface, creating a new window with each agent callback
 * zipcat: zip > base64 > cat for data small/medium data exfil (zstd/x64 or gzip/x86)
 * Remove the use of passing powershell/cmd commands at the main prompt